Add autocomplete functionality for pipeline
