+++ dbt_platform_helper/commands/conduit.py
@@ -12,7 +12,7 @@ from dbt_platform_helper.utils.application import Application
 from dbt_platform_helper.utils.application import load_application
 from dbt_platform_helper.utils.aws import update_postgres_parameter_with_master_secret
 from dbt_platform_helper.utils.click import ClickDocOptCommand
-from dbt_platform_helper.utils.files import is_terraform_project
+from dbt_platform_helper.utils.platform_config import is_terraform_project
 from dbt_platform_helper.utils.versioning import (
     check_platform_helper_version_needs_update,
 )
